# Conflicts:
#	.idea/vcs.xml

+++ app/src/main/java/com/nsu/gomer/barcode2/data/BarcodeRemoteDataSource.kt
@@ -4,6 +4,6 @@ import com.nsu.gomer.barcode2.domain.BarcodeResponse
 import io.reactivex.Single
 
 class BarcodeRemoteDataSource(private val barcodeApi: BarcodeApi): BarcodeDataSource {
-    override fun getBarcodeInfo(code: Int): Single<BarcodeResponse> =
+    override fun getBarcodeInfo(code: Long): Single<BarcodeResponse> =
         barcodeApi.getBarcodeInfo(code)
 }
